Create House Cleaning Service Program
This program calculates the cost for a user who is looking to get a quote on a cleaning service for their home. The user will input the number of small rooms and large rooms when prompted to. The cost for both sized rooms is fixed along with the the tax that is added to the number of rooms being charged for.